+++ lib/wisper/active_record/publisher.rb
@@ -16,19 +16,6 @@ module Wisper
         after_rollback   :after_rollback_broadcast
       end
 
-      def commit(_attributes = nil)
-        warn "[DEPRECATED] use save, create, update_attributes as usual"
-        assign_attributes(_attributes) if _attributes.present?
-        save
-      end
-
-      module ClassMethods
-        def commit(_attributes = nil)
-          warn "[DEPRECATED] use save, create, update_attributes as usual"
-          new(_attributes).save
-        end
-      end
-
       private
 
       def after_validation_broadcast